en proceso

+++ demo/src/main/java/com/example/demo/repositorio/CarritoComprasRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repositorio;
 import com.example.demo.entidades.CarritoCompras;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarritoComprasRepository extends JpaRepository<CarritoCompras, Long> {
+public interface CarritoComprasRepository extends JpaRepository<CarritoCompras, Integer> {
 }
